requirements.txt
python-telegram-bot
currencyapicom
pytest
